% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissIntroJS.R
\name{addIntroJS}
\alias{addIntroJS}
\title{addIntroJS}
\usage{
addIntroJS(inputId = "btnHelp", useVoice = FALSE)
}
\arguments{
\item{inputID}{id for the help button}
}
\description{
Adds introJS to the shiny app. This function should be called in ui.r
This function performs all includes necessary for introJS and puts a green help button on the top right of the screen to start the help.
In server a call should be made to startIntroJS when the button is clicked.
}

