% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissIntroJS.R
\name{dlply.ordered}
\alias{dlply.ordered}
\title{Keeps order while using dlply}
\usage{
dlply.ordered(data, ...)
}
\arguments{
\item{data}{Dataframe}

\item{...}{Parameters passed to dlply}
}
\description{
Keeps order while using dlply
}

